move neural_ent under utils
