--- smt/surrogate_models/neural_net/__init__.py
@@ -1,7 +1,0 @@
-from smt.surrogate_models.neural_net import activation
-from smt.surrogate_models.neural_net import bwd_prop
-from smt.surrogate_models.neural_net import fwd_prop
-from smt.surrogate_models.neural_net import loss
-from smt.surrogate_models.neural_net import model
-from smt.surrogate_models.neural_net import optimizer
-from smt.surrogate_models.neural_net import data
