+++ smt/utils/neural_net/activation.py
